Make "invalid hex" non-fatal
When encountering invalid hex characters in the data section of an URB,
just truncate processing at that point and still output the preceding
data, while printing a warning.

As part of this, imporve the warning to avoid outputing non-printable
characters and include the hex value so that non-printable characters
can be distinguished.